LLM/Chatbot projects/requirements.txt
torch
transformers
pyyaml
tqdm
nltk
rouge-score
